requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
asyncio==3.4.3
PyYAML==5.4.1
